web_app/requirements.txt
flask==2.0.1
transformers==4.30.2
torch==2.0.1
Pillow==9.5.0
numpy==1.24.3
requests==2.31.0
opencv-python==4.8.0.76 